requirements.txt
fastapi
uvicorn[standard]
gunicorn
jinja2
httpx
openai
python-dotenv
motor
azure-search-documents
sse-starlette
slowapi
limits[sqlalchemy]
cryptography
bcrypt
python-jose[cryptography]
python-multipart
apscheduler
azure-monitor-opentelemetry
